Add files via upload
This is a code I wrote to generate a QR code for an event registration forms. 
This code uses simple features of the qrcode library, and explains a little bit how it works, the option to customize your QRcode and how each line works. 
Also include the link to the documentation.